Move util/RankingManager to models/RankingManager
